# public/assets/api.yaml
openapi: 3.0.0
info:
  title: 'My First API'
  version: '0.1'
paths:
...
  /roleUsers:
    get:
      tags:
        - RoleUser
      summary: 'Find list RoleUser'
      description: 'Returns list of RoleUser'
      operationId: getRoleUser
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/RoleUser' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/RoleUser' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/RoleUser' } }
                type: object
        '404':
          description: 'RoleUser not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - RoleUser
      summary: 'Add a new RoleUser to the store'
      operationId: addRoleUser
      requestBody:
        $ref: '#/components/requestBodies/RoleUser'
      responses:
        '201':
          description: 'Created RoleUser'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleUser'
            application/xml:
              schema:
                $ref: '#/components/schemas/RoleUser'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/roleUsers/{id}':
    get:
      tags:
        - RoleUser
      summary: 'Find RoleUser by ID'
      description: 'Returns a single RoleUser'
      operationId: getRoleUserById
      parameters:
        -
          name: id
          in: path
          description: 'ID of RoleUser to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleUser'
            application/xml:
              schema:
                $ref: '#/components/schemas/RoleUser'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'RoleUser not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - RoleUser
      summary: 'Update an existing RoleUser'
      operationId: updateRoleUser
      parameters:
        -
          name: id
          in: path
          description: 'RoleUser id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/RoleUser'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'RoleUser not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - RoleUser
      summary: 'Deletes a RoleUser'
      operationId: deleteRoleUser
      parameters:
        -
          name: id
          in: path
          description: 'RoleUser id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /roles:
    get:
      tags:
        - Roles
      summary: 'Find list Roles'
      description: 'Returns list of Roles'
      operationId: getRoles
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Roles' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Roles' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Roles' } }
                type: object
        '404':
          description: 'Roles not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Roles
      summary: 'Add a new Roles to the store'
      operationId: addRoles
      requestBody:
        $ref: '#/components/requestBodies/Roles'
      responses:
        '201':
          description: 'Created Roles'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roles'
            application/xml:
              schema:
                $ref: '#/components/schemas/Roles'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/roles/{id}':
    get:
      tags:
        - Roles
      summary: 'Find Roles by ID'
      description: 'Returns a single Roles'
      operationId: getRolesById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Roles to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roles'
            application/xml:
              schema:
                $ref: '#/components/schemas/Roles'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Roles not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Roles
      summary: 'Update an existing Roles'
      operationId: updateRoles
      parameters:
        -
          name: id
          in: path
          description: 'Roles id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Roles'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Roles not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Roles
      summary: 'Deletes a Roles'
      operationId: deleteRoles
      parameters:
        -
          name: id
          in: path
          description: 'Roles id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    RoleUser:
      title: RoleUser
      description: RoleUser
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        user_id:
          title: user_id
          description: user_id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        role_id:
          title: role_id
          description: role_id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        created_at:
          title: created_at
          description: created_at
          type: string
          format: '-'
          nullable: true
        updated_at:
          title: updated_at
          description: updated_at
          type: string
          format: '-'
          nullable: true
      type: object
    Roles:
      title: Roles
      description: Roles
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        name:
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        created_at:
          title: created_at
          description: created_at
          type: string
          format: '-'
          nullable: true
        updated_at:
          title: updated_at
          description: updated_at
          type: string
          format: '-'
          nullable: true
      type: object
...
  requestBodies:
...
    RoleUser:
      description: 'RoleUser object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/RoleUser'
        application/xml:
          schema:
            $ref: '#/components/schemas/RoleUser'
        application/json:
          schema:
            $ref: '#/components/schemas/RoleUser'
    Roles:
      description: 'Roles object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Roles'
        application/xml:
          schema:
            $ref: '#/components/schemas/Roles'
        application/json:
          schema:
            $ref: '#/components/schemas/Roles'
...